test: split tests for syncServices and syncSubscription functions into two separate files
